ai-visuzalize-services/requirements.txt
Flask==3.0.0
greenlet==3.0.3
numpy
opencv_python
PyYAML==6.0.1
redis==5.0.1
simplejpeg
gunicorn
python-dotenv==1.0.1
